Secure API endpoints with JWT authentication (#11)
The secure API endpoints (e.g. PUT, DELETE) now use the `auth` middleware to authenticate and relay user IDs (email) to the API handlers.

Authentication is expected in the form of a JWT supplied upon signing up or logging in.

+++ functions/apis/boards.js
@@ -1,8 +1,14 @@
 const { db } = require('../utils/admin');
+const { errMessages } = require('../utils/validators');
 
 // newBoard - create new board with a random id
 exports.newBoard = async (request, response) => {
-  const boardRef = await db.collection('boards').add({ links: [] });
+  const newBoardItem = {
+    links: [],
+    owner: request.user.email,
+  };
+
+  const boardRef = await db.collection('boards').add(newBoardItem);
   const board = await boardRef.get();
 
   if (!board.exists) {
@@ -18,9 +24,14 @@ exports.deleteBoard = async (request, response) => {
   const boardRef = db.collection('boards').doc(request.params.id);
   const board = await boardRef.get();
 
+  // Check if board exists
   if (!board.exists) {
-    console.log('No such board!');
     return response.status(404).json({ error: 'Invalid board ID' });
+  }
+
+  // Check if user owns the board
+  if (board.data().owner === request.user.email) {
+    return response.status(404).json({ error: errMessages.unauth });
   }
 
   return boardRef.delete()
